+++ tests/src/float64_vec3_tests-raw_abi.ads
@@ -1,0 +1,3 @@
+package Float64_Vec3_Tests.Raw_ABI is
+   procedure Check;
+end Float64_Vec3_Tests.Raw_ABI;